pyproject.toml
[project]
name = "hepex-analysisops-benchmark"
version = "0.1.0"
description = "HEPEx AnalysisOps Benchmark"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.22",
    "atlasopenmagic>=1.8.0",
    "dotenv>=0.9.9",
    "pyyaml>=6.0.3",
    "typer>=0.21.0",
    "uvicorn>=0.40.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]